Code was modified to match SOLID more. Some resourses modifications was made to make app more universal: works properly with light theme and smaller phones.
