+++ app/src/main/java/com/example/students_camp_practise/assets/MessagesData.kt
@@ -1,0 +1,37 @@
+package com.example.students_camp_practise.assets
+
+import com.example.students_camp_practise.R
+import com.example.students_camp_practise.data.Message
+
+object SampleData { //object with list of user reviews information (Message data class)
+    val feedbackSample = listOf(
+        Message(
+            "Marius Conte",
+            "February 14, 2019",
+            "Once you start to learn its secrets, there’s a wild and exciting variety of play here that’s unmatched, even by its peers.",
+            R.drawable.user_logo
+        ),
+        Message(
+            "Maria Marcelino",
+            "November 12, 2017",
+            "List of Android versions:\n" +
+                    "Android KitKat (API 19)\n" +
+                    "Android Lollipop (API 21)\n" +
+                    "Android Marshmallow (API 23)\n" +
+                    "Android Nougat (API 24)\n" +
+                    "Android Oreo (API 26)\n" +
+                    "Android Pie (API 28)\n" +
+                    "Android 10 (API 29)\n" +
+                    "Android 11 (API 30)\n" +
+                    "Android 12 (API 31)\n",
+            R.drawable.user_logo_2
+        ),
+        Message(
+            "Lisa Ajax",
+            "February 5, 2016",
+            "I think Kotlin will be my favorite programming language.\n" +
+                    "It's so much fun!",
+            R.drawable.user_logo_3
+        ),
+    )
+}
