+++ app/src/main/java/com/example/students_camp_practise/data/Message.kt
@@ -1,0 +1,8 @@
+package com.example.students_camp_practise.data
+
+data class Message(
+    val author: String,
+    val time: String,
+    val body: String,
+    val uLogo:Int
+)
