+++ app/src/main/java/com/example/students_camp_practise/data/Tag.kt
@@ -1,0 +1,6 @@
+package com.example.students_camp_practise.data
+
+data class Tag(
+    val tagName:String,
+    val onClickMessage: TagMessage
+)
