+++ app/src/main/java/com/example/students_camp_practise/data/TagMessage.kt
@@ -1,0 +1,6 @@
+package com.example.students_camp_practise.data
+
+data class TagMessage(
+    val name: String,
+    val category: String
+)
